# api-reference/static/openapi.yaml
openapi: 3.1.0
info:
  title: Samantha API
  version: 1.0.0
  description: |
    The Samantha external API surface provides programmatic access to portfolio, work-item, topology, and admin features.

    All responses on error return RFC 9457 Problem Details format.
    Authentication uses Bearer token via Authorization header.
  contact:
    name: Samantha API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:5100/samantha/v1
    description: Local development
  - url: https://api.example.com/samantha/v1
    description: Production
...
paths:
...
  /auth/password-reset:
    post:
      tags: [auth]
      summary: Request password reset
      operationId: passwordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required: [email]
      responses:
        '204':
          description: Reset email sent
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Confirm password reset
      operationId: passwordResetConfirm
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                newPassword:
                  type: string
              required: [token, newPassword]
      responses:
        '204':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
              required: [currentPassword, newPassword]
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type:
          type: string
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        violations:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Bad Request (400)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized (401)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT